chore!: drop old realtime client
