--- Sources/Realtime/Deprecated.swift
@@ -1,11 +1,0 @@
-//
-//  Deprecated.swift
-//
-//
-//  Created by Guilherme Souza on 23/12/23.
-//
-
-import Foundation
-
-@available(*, deprecated, renamed: "RealtimeMessage")
-public typealias Message = RealtimeMessage
