+++ Sources/Supabase/SupabaseClient.swift
@@ -66,10 +66,11 @@ public final class SupabaseClient: Sendable {
     }
   }
 
-  let _realtime: UncheckedSendable<RealtimeClient>
-
   /// Realtime client for Supabase
-  public let realtimeV2: RealtimeClientV2
+  public let realtime: RealtimeClient
+
+  @available(*, deprecated, renamed: "realtime")
+  public var realtimeV2: RealtimeClient { realtime }
 
   /// Supabase Functions allows you to deploy and invoke edge functions.
   public var functions: FunctionsClient {
@@ -170,14 +171,6 @@ public final class SupabaseClient: Sendable {
       autoRefreshToken: options.auth.autoRefreshToken
     )
 
-    _realtime = UncheckedSendable(
-      RealtimeClient(
-        supabaseURL.appendingPathComponent("/realtime/v1").absoluteString,
-        headers: _headers.dictionary,
-        params: _headers.dictionary
-      )
-    )
-
     var realtimeOptions = options.realtime
     realtimeOptions.headers.merge(with: _headers)
 
@@ -185,7 +178,7 @@ public final class SupabaseClient: Sendable {
       realtimeOptions.logger = options.global.logger
     }
 
-    realtimeV2 = RealtimeClientV2(
+    realtime = RealtimeClient(
       url: supabaseURL.appendingPathComponent("/realtime/v1"),
       options: realtimeOptions
     )
@@ -239,10 +232,8 @@ public final class SupabaseClient: Sendable {
   }
 
   /// Returns all Realtime channels.
-  public var channels: [RealtimeChannelV2] {
-    get async {
-      await Array(realtimeV2.subscriptions.values)
-    }
+  public var channels: [RealtimeChannel] {
+    Array(realtime.subscriptions.values)
   }
 
   /// Creates a Realtime channel with Broadcast, Presence, and Postgres Changes.
@@ -252,19 +243,19 @@ public final class SupabaseClient: Sendable {
   public func channel(
     _ name: String,
     options: @Sendable (inout RealtimeChannelConfig) -> Void = { _ in }
-  ) async -> RealtimeChannelV2 {
-    await realtimeV2.channel(name, options: options)
+  ) -> RealtimeChannel {
+    realtime.channel(name, options: options)
   }
 
   /// Unsubscribes and removes Realtime channel from Realtime client.
   /// - Parameter channel: The Realtime channel to remove.
-  public func removeChannel(_ channel: RealtimeChannelV2) async {
-    await realtimeV2.removeChannel(channel)
+  public func removeChannel(_ channel: RealtimeChannel) async {
+    await realtime.removeChannel(channel)
   }
 
   /// Unsubscribes and removes all Realtime channels from Realtime client.
   public func removeAllChannels() async {
-    await realtimeV2.removeAllChannels()
+    await realtime.removeAllChannels()
   }
 
   /// Handles an incoming URL received by the app.
@@ -373,7 +364,6 @@ public final class SupabaseClient: Sendable {
       return nil
     }
 
-    realtime.setAuth(accessToken)
-    await realtimeV2.setAuth(accessToken)
+    await realtime.setAuth(accessToken)
   }
 }
