+++ Tests/RealtimeTests/CallbackManagerTests.swift
@@ -130,7 +130,7 @@ final class CallbackManagerTests: XCTestCase {
       commitTimestamp: currentDate,
       record: ["email": .string("new@mail.com")],
       oldRecord: ["email": .string("old@mail.com")],
-      rawMessage: RealtimeMessageV2(joinRef: nil, ref: nil, topic: "", event: "", payload: [:])
+      rawMessage: RealtimeMessage(joinRef: nil, ref: nil, topic: "", event: "", payload: [:])
     )
     callbackManager.triggerPostgresChanges(ids: [updateUsersId], data: .update(updateUserAction))
 
@@ -138,7 +138,7 @@ final class CallbackManagerTests: XCTestCase {
       columns: [],
       commitTimestamp: currentDate,
       record: ["email": .string("email@mail.com")],
-      rawMessage: RealtimeMessageV2(joinRef: nil, ref: nil, topic: "", event: "", payload: [:])
+      rawMessage: RealtimeMessage(joinRef: nil, ref: nil, topic: "", event: "", payload: [:])
     )
     callbackManager.triggerPostgresChanges(ids: [insertUsersId], data: .insert(insertUserAction))
 
@@ -149,7 +149,7 @@ final class CallbackManagerTests: XCTestCase {
       columns: [],
       commitTimestamp: currentDate,
       oldRecord: ["id": .string("1234")],
-      rawMessage: RealtimeMessageV2(joinRef: nil, ref: nil, topic: "", event: "", payload: [:])
+      rawMessage: RealtimeMessage(joinRef: nil, ref: nil, topic: "", event: "", payload: [:])
     )
     callbackManager.triggerPostgresChanges(
       ids: [deleteSpecificUserId],
@@ -174,7 +174,7 @@ final class CallbackManagerTests: XCTestCase {
     XCTAssertNoLeak(callbackManager)
 
     let event = "new_user"
-    let message = RealtimeMessageV2(
+    let message = RealtimeMessage(
       joinRef: nil,
       ref: nil,
       topic: "realtime:users",
@@ -209,7 +209,7 @@ final class CallbackManagerTests: XCTestCase {
     callbackManager.triggerPresenceDiffs(
       joins: joins,
       leaves: leaves,
-      rawMessage: RealtimeMessageV2(joinRef: nil, ref: nil, topic: "", event: "", payload: [:])
+      rawMessage: RealtimeMessage(joinRef: nil, ref: nil, topic: "", event: "", payload: [:])
     )
 
     XCTAssertNoDifference(receivedAction.value?.joins, joins)
