+++ Tests/RealtimeTests/_PushTests.swift
@@ -12,7 +12,7 @@ import XCTest
 
 final class _PushTests: XCTestCase {
   var ws: MockWebSocketClient!
-  var socket: RealtimeClientV2!
+  var socket: RealtimeClient!
 
   override func invokeTest() {
     withMainSerialExecutor {
@@ -24,7 +24,7 @@ final class _PushTests: XCTestCase {
     super.setUp()
 
     ws = MockWebSocketClient()
-    socket = RealtimeClientV2(
+    socket = RealtimeClient(
       url: URL(string: "https://localhost:54321/v1/realtime")!,
       options: RealtimeClientOptions(
         headers: ["apiKey": "apikey"]
@@ -34,7 +34,7 @@ final class _PushTests: XCTestCase {
   }
 
   func testPushWithoutAck() async {
-    let channel = RealtimeChannelV2(
+    let channel = RealtimeChannel(
       topic: "realtime:users",
       config: RealtimeChannelConfig(
         broadcast: .init(acknowledgeBroadcasts: false),
@@ -44,9 +44,9 @@ final class _PushTests: XCTestCase {
       socket: Socket(client: socket),
       logger: nil
     )
-    let push = PushV2(
+    let push = Push(
       channel: channel,
-      message: RealtimeMessageV2(
+      message: RealtimeMessage(
         joinRef: nil,
         ref: "1",
         topic: "realtime:users",
@@ -61,7 +61,7 @@ final class _PushTests: XCTestCase {
 
   // FIXME: Flaky test, it fails some time due the task scheduling, even tho we're using withMainSerialExecutor.
 //  func testPushWithAck() async {
-//    let channel = RealtimeChannelV2(
+//    let channel = RealtimeChannel(
 //      topic: "realtime:users",
 //      config: RealtimeChannelConfig(
 //        broadcast: .init(acknowledgeBroadcasts: true),
@@ -70,9 +70,9 @@ final class _PushTests: XCTestCase {
 //      socket: socket,
 //      logger: nil
 //    )
-//    let push = PushV2(
+//    let push = Push(
 //      channel: channel,
-//      message: RealtimeMessageV2(
+//      message: RealtimeMessage(
 //        joinRef: nil,
 //        ref: "1",
 //        topic: "realtime:users",
